\name{ExpMet-class}
\Rdversion{1.1}
\docType{class}
\alias{ExpMet-class}

\title{Class \code{"ExpMet"}}
\description{
This S4 class holds the names of all libraries used in experiment. It contains a slot of character vector called ExpName
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("ExpMet", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{ExpName}:}{Object of class \code{"character"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "ExpMet" in the signature.
}
\references{
%%  ~~put references to the literature/web site here~~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
}
\examples{
showClass("ExpMet")
}
\keyword{classes}
